Rewrote Navigation using tailwind
