+++ src/app/layout.module.scss
@@ -9,7 +9,4 @@
   & > main {
     flex: 1;
   }
-  & > * {
-    display: block;
-  }
 }
