--- src/components/organisms/Navigation/index.module.scss
@@ -1,70 +1,0 @@
-@use '@/styles/mixins';
-
-#wide_nav {
-  background: #66a928d3;
-  text-align: center;
-  z-index: 50;
-
-  //position: absolute;
-  display: flex;
-  align-items: flex-end;
-  width: 100%;
-  height: var(--navigation-height);
-
-  a {
-    display: inline-block;
-    font-family: var(--font-comfortaa);
-    text-decoration: none;
-    color: white;
-    padding: 1em;
-  }
-
-  box-shadow: 0 3px 0 #54a60c;
-  @include mixins.mq(dark) {
-    box-shadow: 0 3px 0 #27341a;
-  }
-}
-
-#wide_nav_wrapper {
-  width: 100%;
-  vertical-align: bottom;
-  margin-bottom: 10px;
-}
-
-@include mixins.mq(dark) {
-  #wide_nav {
-    background: #2d4b11d3;
-  }
-}
-
-@include mixins.mq(pc) {
-  .side_menu_link {
-    border-radius: 1000px;
-    height: 1.3em;
-    margin: 0 2px;
-    padding: 6px 12px 0 !important;
-    transition: 0.1s;
-
-    &[data-current-page='true'] {
-      background: white;
-      color: #66a928d3 !important;
-      @include mixins.mq(dark) {
-        color: #2d4b11d3 !important;
-      }
-      cursor: default;
-    }
-
-    &[data-current-page='false']:hover {
-      background: rgba(255, 255, 255, 0.2);
-      a:hover {
-        color: #90e200;
-      }
-    }
-  }
-}
-
-@include mixins.mq(sp) {
-  #wide_nav {
-    display: none;
-  }
-}
