+++ src/styles/globals.scss
@@ -80,9 +80,6 @@ body {
   outline-color: #90e200;
 }
 
-main {
-  letter-spacing: 0.03em;
-}
 .good-break {
   overflow-wrap: anywhere;
   word-break: keep-all;
